a11y for token views https://github.com/xch-dev/sage/issues/275
